liveness: move TimeUntilStoreDead to liveness
Previously this was in the storepool package, however other places in
the code had similar decisions based on their own notion of dead/alive.
This begins to centralize that logic.

Epic: none

Release note: None
