+++ pkg/kv/kvserver/allocator/storepool/store_pool_test.go
@@ -19,6 +19,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/liveness"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/liveness/livenesspb"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
@@ -54,7 +55,7 @@ func TestStorePoolGossipUpdate(t *testing.T) {
 	ctx := context.Background()
 	st := cluster.MakeTestingClusterSettings()
 	stopper, g, _, sp, _ := CreateTestStorePool(ctx, st,
-		TestTimeUntilStoreDead, false, /* deterministic */
+		liveness.TestTimeUntilStoreDead, false, /* deterministic */
 		func() int { return 0 }, /* NodeCount */
 		livenesspb.NodeLivenessStatus_DEAD)
 	defer stopper.Stop(ctx)
@@ -124,7 +125,7 @@ func TestStorePoolGetStoreList(t *testing.T) {
 	st := cluster.MakeTestingClusterSettings()
 	// We're going to manually mark stores dead in this test.
 	stopper, g, _, sp, mnl := CreateTestStorePool(ctx, st,
-		TestTimeUntilStoreDead, false, /* deterministic */
+		liveness.TestTimeUntilStoreDead, false, /* deterministic */
 		func() int { return 10 }, /* nodeCount */
 		livenesspb.NodeLivenessStatus_DEAD)
 	defer stopper.Stop(ctx)
@@ -420,7 +421,7 @@ func TestStorePoolGetStoreDetails(t *testing.T) {
 	ctx := context.Background()
 	st := cluster.MakeTestingClusterSettings()
 	stopper, g, _, sp, _ := CreateTestStorePool(ctx, st,
-		TestTimeUntilStoreDead, false, /* deterministic */
+		liveness.TestTimeUntilStoreDead, false, /* deterministic */
 		func() int { return 10 }, /* nodeCount */
 		livenesspb.NodeLivenessStatus_DEAD)
 	defer stopper.Stop(ctx)
@@ -443,7 +444,7 @@ func TestStorePoolFindDeadReplicas(t *testing.T) {
 	ctx := context.Background()
 	st := cluster.MakeTestingClusterSettings()
 	stopper, g, _, sp, mnl := CreateTestStorePool(ctx, st,
-		TestTimeUntilStoreDead, false, /* deterministic */
+		liveness.TestTimeUntilStoreDead, false, /* deterministic */
 		func() int { return 10 }, /* nodeCount */
 		livenesspb.NodeLivenessStatus_DEAD)
 	defer stopper.Stop(ctx)
@@ -549,7 +550,7 @@ func TestStorePoolDefaultState(t *testing.T) {
 	ctx := context.Background()
 	st := cluster.MakeTestingClusterSettings()
 	stopper, _, _, sp, _ := CreateTestStorePool(ctx, st,
-		TestTimeUntilStoreDead, false, /* deterministic */
+		liveness.TestTimeUntilStoreDead, false, /* deterministic */
 		func() int { return 10 }, /* nodeCount */
 		livenesspb.NodeLivenessStatus_DEAD)
 	defer stopper.Stop(ctx)
@@ -580,7 +581,7 @@ func TestStorePoolThrottle(t *testing.T) {
 	ctx := context.Background()
 	st := cluster.MakeTestingClusterSettings()
 	stopper, g, _, sp, _ := CreateTestStorePool(ctx, st,
-		TestTimeUntilStoreDead, false, /* deterministic */
+		liveness.TestTimeUntilStoreDead, false, /* deterministic */
 		func() int { return 10 }, /* nodeCount */
 		livenesspb.NodeLivenessStatus_DEAD)
 	defer stopper.Stop(ctx)
@@ -608,13 +609,13 @@ func TestStorePoolSuspected(t *testing.T) {
 	ctx := context.Background()
 	st := cluster.MakeTestingClusterSettings()
 	stopper, g, _, sp, mnl := CreateTestStorePool(ctx, st,
-		TestTimeUntilStoreDeadOff, false, /* deterministic */
+		liveness.TestTimeUntilStoreDeadOff, false, /* deterministic */
 		func() int { return 10 }, /* nodeCount */
 		livenesspb.NodeLivenessStatus_DEAD)
 	defer stopper.Stop(ctx)
 
 	now := sp.clock.Now()
-	timeUntilStoreDead := TimeUntilStoreDead.Get(&sp.st.SV)
+	timeUntilStoreDead := liveness.TimeUntilStoreDead.Get(&sp.st.SV)
 	timeAfterStoreSuspect := TimeAfterStoreSuspect.Get(&sp.st.SV)
 
 	// Verify a store that we haven't seen yet is unknown status.
@@ -691,7 +692,7 @@ func TestGetLocalities(t *testing.T) {
 	ctx := context.Background()
 	st := cluster.MakeTestingClusterSettings()
 	stopper, g, _, sp, _ := CreateTestStorePool(ctx, st,
-		TestTimeUntilStoreDead, false, /* deterministic */
+		liveness.TestTimeUntilStoreDead, false, /* deterministic */
 		func() int { return 10 }, /* nodeCount */
 		livenesspb.NodeLivenessStatus_DEAD)
 	defer stopper.Stop(ctx)
@@ -773,7 +774,7 @@ func TestStorePoolDecommissioningReplicas(t *testing.T) {
 	ctx := context.Background()
 	st := cluster.MakeTestingClusterSettings()
 	stopper, g, _, sp, mnl := CreateTestStorePool(ctx, st,
-		TestTimeUntilStoreDead, false, /* deterministic */
+		liveness.TestTimeUntilStoreDead, false, /* deterministic */
 		func() int { return 10 }, /* nodeCount */
 		livenesspb.NodeLivenessStatus_DEAD)
 	defer stopper.Stop(ctx)
