+++ pkg/kv/kvserver/store_pool_test.go
@@ -16,6 +16,7 @@ import (
 	"time"
 
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/allocator/storepool"
+	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/liveness"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/liveness/livenesspb"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/load"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/logstore"
@@ -44,7 +45,7 @@ func TestStorePoolUpdateLocalStore(t *testing.T) {
 	// We're going to manually mark stores dead in this test.
 	st := cluster.MakeTestingClusterSettings()
 	stopper, g, _, sp, _ := storepool.CreateTestStorePool(ctx, st,
-		storepool.TestTimeUntilStoreDead, false, /* deterministic */
+		liveness.TestTimeUntilStoreDead, false, /* deterministic */
 		func() int { return 10 }, /* nodeCount */
 		livenesspb.NodeLivenessStatus_DEAD)
 	defer stopper.Stop(ctx)
@@ -216,7 +217,7 @@ func TestStorePoolUpdateLocalStoreBeforeGossip(t *testing.T) {
 	cfg := TestStoreConfig(clock)
 	var stopper *stop.Stopper
 	stopper, _, _, cfg.StorePool, _ = storepool.CreateTestStorePool(ctx, cfg.Settings,
-		storepool.TestTimeUntilStoreDead, false, /* deterministic */
+		liveness.TestTimeUntilStoreDead, false, /* deterministic */
 		func() int { return 10 }, /* nodeCount */
 		livenesspb.NodeLivenessStatus_DEAD)
 	defer stopper.Stop(ctx)
